Clean up Kalman demo code
