+++ src/articles/components/kalman-filter.tsx
@@ -16,9 +16,69 @@ import {
   subtract
 } from 'mathjs';
 import { MOBILE } from '../../constants';
+import LoadImage from './load';
+import ChartBackground from './chart-background';
 import defaultMapping from './kalman-mapping.json';
 import defaultParams from './kalman-params.json';
 import defaultInput from './kalman-input.json';
+
+//
+// Kalman Filter in JavaScript
+//
+
+const kalmanFilter = ({
+  inputs,
+  measurements,
+  A, B, C, D, P0, Q, R, x0
+}) => {
+  const Atrans = transpose(A);
+  const Ctrans = transpose(C);
+  const I = identity(x0.size());
+
+  let y = measurements[0];
+  let x = x0;
+  let P = P0;
+
+  return measurements.map((z, index) => {
+    // Input
+    const u = inputs[index];
+
+    // Update covariance
+    P = add(multiply(multiply(A, P), Atrans), Q);
+
+    // Optimize gain
+    const K = divide(
+      multiply(P, Ctrans),
+      add(
+        multiply(multiply(C, P), Ctrans),
+        R
+      ));
+
+    // Correct state with measurement
+    x = add(x, multiply(K, (z - y)));
+
+    // Correct covariance
+    P = add(
+      multiply(
+        multiply(subtract(I, multiply(K, C)), P),
+        transpose(subtract(I, multiply(K, C)))),
+      multiply(multiply(K, R), transpose(K))
+    );
+
+    // Predict
+    y = add(multiply(C, x), multiply(D, u))._data[0][0];
+
+    // Update state
+    x = add(multiply(A, x), multiply(B, u));
+
+    // Output
+    return y;
+  });
+};
+
+//
+// Styled Components
+//
 
 const Wrapper = styled.section`
   display: grid;
@@ -63,45 +123,6 @@ const Params = styled.div`
   display: flex;
   flex-direction: column;
 `;
-
-const LoadImage = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="65.4px"
-    height="81.7px"
-    viewBox="0 0 102.2 127.6"
-  >
-    <path fill="#CDCDCD" d="M0.7,39c0-6.6,0-13.1,0-19.7c0-3.2-0.2-6.4,0-9.5c0.1-2.7,1.5-5.2,3.7-6.7c2.7-1.7,5.4-2,8.5-1.9
-      c3.1,0.1,6.2,0.1,9.3,0.1c6.4-0.1,12.8,0,19.3-0.1c12.8-0.2,25.7,0,38.6,0c3.2,0,6.5,0,9.7,0c2.8,0,5.5,0,7.8,1.6
-      c2,1.4,3.5,3.4,3.9,5.9c0.2,1.5,0.1,3,0.1,4.5c0,1.7,0,3.3,0,5c0,13,0.3,26,0.3,39c0,3.1,0,6.1,0,9.2c0,0.7,0.4,3.6-1,3
-      c-1.4-0.6-2.7-1.6-4-2.4c-0.7-0.4-1.3-0.8-2-1.2c-0.5-0.3-1.4-0.8-1.8-1.3c-0.3-0.5-0.2-1.4-0.2-2c0-0.9-0.1-1.7-0.1-2.6
-      c-0.2-3.2-0.5-6.4-0.7-9.7c-0.1-1.5-0.1-3.1-0.4-4.6c-0.3-1.4-0.9-2.8-1.6-4c-0.7-1.2-1.5-2.2-2.4-3.1c-1-0.9-2-2-3.1-2.6
-      c-2.5-1.5-5.7-2-8.6-2c-3.2,0-6.5,0-9.7,0c-12.8,0-25.7-0.2-38.5,0.1c-3.1,0.1-5.9,0-8.8,1.3c-2.6,1.2-4.8,3.1-6.3,5.5
-      c-1.7,2.6-2.2,5.5-2.4,8.6c-0.2,3.2-0.3,6.4-0.6,9.7c-0.1,1.5,0,3.2-0.2,4.7c-0.2,1.2-1.6,1.7-2.5,2.2c-1.3,0.8-2.6,1.6-3.9,2.4
-      c-0.4,0.3-1.3,1-1.8,1c-0.7,0-0.6-0.8-0.6-1.3c0-6.5,0.1-13.1,0-19.6C0.7,45.2,0.7,42.1,0.7,39C0.7,39,0.7,39,0.7,39z"/>
-    <path fill="#CDCDCD" d="M96,107.5c-0.4,3.3,0,6.7-0.1,10c0,3-1.3,5.9-4,7.4c-2.4,1.4-5.7,1.7-8.2,0.4c-2.6-1.3-4.3-3.9-4.5-6.8
-      c-0.3-3.7,0-7.4-0.2-11.1c-11.5-0.4-23,0-34.5-0.1c-5.7-0.1-11.4-0.1-17.1,0c-0.6,0-1.3,0-1.9,0c-0.6,0-1.4-0.1-1.9,0.1
-      c-0.7,0.3-0.4,1.9-0.4,2.5c0,2.9,0.2,5.8,0,8.7c-0.3,4.6-4.9,7.9-9.3,7.4c-2.2-0.2-4.4-1.4-5.8-3.2c-1.7-2.2-1.6-4.8-1.6-7.4
-      c0-1.4,0-2.9,0-4.3c0-0.7,0-1.3,0-2c0-0.4,0.1-1,0-1.4c-0.2-0.7-1.4-0.4-1.9-0.4c-0.6,0-1.3,0-1.9,0c-0.4,0-1.1,0.1-1.5-0.1
-      c-0.7-0.4-0.4-2.6-0.4-3.2c0-1.4,0-2.8,0-4.1c0.1-2.8,0-5.6,0-8.4c0-2.9,0-5.8-0.1-8.7c0-1.5,0-2.9,0-4.4c0-1-0.4-2.6,0.3-3.5
-      c0.7-0.9,2.2-1.5,3.2-2.1c1.2-0.7,2.3-1.4,3.5-2.1c1.2-0.7,2.4-1.5,3.6-2.2c1-0.6,2.2-1.1,2.4-2.4c0.2-1.4,0.1-2.9,0.1-4.4
-      c0-1.4,0.1-2.8,0.2-4.2c0.2-2.9,0.4-5.7,0.6-8.6c0.2-2.7,0.9-5.1,2.8-7.1c1.8-1.9,4.1-3,6.7-3.3c1.4-0.2,2.9-0.1,4.3,0
-      c1.4,0,2.8,0,4.2,0c2.9,0,5.9,0,8.8,0c5.7,0,11.4,0,17.1,0c5.9,0,11.7-0.1,17.6,0c2.6,0,5.1,0.6,7.2,2.2c1.8,1.4,3.6,4,3.9,6.3
-      c0.6,5.7,0.9,11.5,1.3,17.2c0.1,1-0.1,2.5,0.7,3.3c0.5,0.4,1.1,0.7,1.7,1c0.6,0.4,1.2,0.7,1.8,1.1c2.3,1.4,4.7,2.8,7,4.2
-      c0.5,0.3,1.3,0.7,1.6,1.1c0.4,0.4,0.3,1.1,0.4,1.6c0.1,1.4,0,2.8,0,4.2c0,2.8,0.1,5.6,0.1,8.5c0,2.9-0.2,5.7-0.1,8.6
-      c0.1,2.9,0,5.7-0.1,8.6c0,1.3-0.9,1.2-2,1.2C98.5,107.4,97.3,107.4,96,107.5z M72.8,66.9c2.9,0,5.7,0,8.6,0.1c0.8,0,2.3,0.3,2.4-0.8
-      c0.1-1.3-0.1-2.8-0.2-4.1c-0.2-2.9-0.4-5.7-0.6-8.5c-0.1-1.4-0.3-2.9-0.3-4.3c0-1.2-0.2-2.3-0.9-3.3c-0.6-0.9-1.4-1.7-2.4-2.2
-      c-1.1-0.5-2.4-0.5-3.6-0.5c-2.9,0-5.8,0-8.7,0c-11.7-0.1-23.4,0-35.1,0c-1.4,0-2.9,0-4.3,0c-1.2,0-2.6-0.1-3.8,0.3
-      c-2.5,0.8-3.7,3.2-4,5.6c-0.2,2.8-0.3,5.6-0.5,8.5c-0.1,1.5-0.2,3-0.3,4.5c-0.1,1.3-0.4,2.8-0.2,4c0.2,1,1.3,0.8,2.1,0.8
-      c1.4-0.1,2.8,0,4.2,0C28.1,66.9,67.1,66.8,72.8,66.9z M35.8,75.8c-1.2,0-2.4,0-3.6,0c-1,0-1.9,0-2.8,0.6c-1.9,1.3-1.3,4-1.3,5.9
-      c0,2.3,0,4.5,0,6.8c0,1.1,0,2.3,0,3.4c0,1-0.2,2.3,0.1,3.2c0.3,0.8,1.1,0.7,1.8,0.7c1.1,0,2.1,0,3.2,0c2.3,0,4.5,0,6.8,0
-      c4.5,0,8.9,0,13.4,0.1c4.4,0,8.8-0.1,13.2,0c2.1,0,4.4,0.3,6.5,0c1.1-0.2,1.3-0.4,1.3-1.4c0-1,0-2,0-3c0-4.4,0.1-8.9-0.1-13.3
-      c-0.1-1.7-1.5-2.9-3.2-3.1C69,75.6,40.9,75.8,35.8,75.8z M88,78.9c-1.7,0.2-3,0.5-4.4,1.7c-1.3,1.1-1.9,2.3-2.2,4
-      c-0.4,2.7,0.7,5.2,3,6.7c2.2,1.5,5.1,1.4,7.3-0.1c2.4-1.7,3.4-4.7,2.5-7.6C93.4,80.9,90.9,79,88,78.9z M20.9,85.9
-      c0.1-1.7-0.4-3-1.5-4.3c-1-1.1-2.2-1.9-3.6-2.3c-2.5-0.8-5.3,0.4-6.8,2.5c-1.8,2.4-1.6,5.9,0.3,8.2c0.8,1,1.8,1.7,2.9,2.2
-      c1.5,0.6,2.8,0.6,4.2,0C19,91.2,21.2,88.8,20.9,85.9z"/>
-  </svg>
-)
 
 const Load = styled.div`
   display: flex;
@@ -223,6 +244,10 @@ const ChartArea = styled.section`
   margin: 24px 16px 0 16px;
 `;
 
+//
+// Helpers
+//
+
 const getPoints = (samples, min, max, width, height, xOffset, yOffset) => {
   const offset = min;
   const factor = (height - 2) / (max - min);
@@ -235,336 +260,6 @@ const getPoints = (samples, min, max, width, height, xOffset, yOffset) => {
       return `${x},${y}`;
     })
     ?.join(' ') ?? '';
-};
-
-const Chart = ({
-  rows,
-  columnIndex,
-  outputs,
-  width
-}) => {
-  const height = 200;
-  const rulerMarkSize = 40;
-
-  const samples = useMemo(
-    () => rows.map(r => Number(r[columnIndex])),
-  [rows, columnIndex]);
-  
-  const { min, max } = useMemo(() => samples.reduce(
-    ({ min, max }, sample) => ({
-      min: min === undefined ? sample : Math.min(min, sample),
-      max: max === undefined ? sample : Math.max(max, sample)
-    }), {}),
-  [samples]);
-
-  if (!width) return null;
-
-  return (
-    <svg
-      width={`${width}px`}
-      height={`${height}px`}
-      viewBox={`0 0 ${width} ${height}`}
-    >
-      <circle fill="#888888" cx="42.7" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="68.6" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="47.8" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="26.9" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="6" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="42.7" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="63.6" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="84.5" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="105.4" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="126.3" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="147.1" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="168" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="188.9" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="209.8" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="230.7" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="251.5" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="272.4" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="293.3" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="314.2" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="335.1" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="355.9" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="376.8" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="397.7" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="418.6" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="439.5" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="460.3" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="481.2" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="502.1" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="523" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="543.9" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="564.7" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="173" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="152.2" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="131.3" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="110.4" r="1.4"/>
-      <circle fill="#888888" cx="585.6" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="606.5" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="627.4" cy="89.5" r="1.4"/>
-      <circle fill="#888888" cx="648.3" cy="89.5" r="1.4"/>
-      <polyline
-        fill="none"
-        stroke="#FF008E"
-        strokeMiterlimit="10"
-        points={getPoints(
-          samples, min, max, width - rulerMarkSize, height - rulerMarkSize, rulerMarkSize, 0
-        )}
-      />
-      <polyline
-        fill="none"
-        stroke="#12C0E1"
-        strokeMiterlimit="10"
-        points={getPoints(
-          outputs, min, max, width - rulerMarkSize, height - rulerMarkSize, rulerMarkSize, 0
-        )}
-      />
-      <g>
-        <g>
-          <polyline strokeWidth="1.5" fill="none" stroke="#AAAAAA" points="1.2,5.4 25.5,5.4 25.5,168.2 1.2,168.2 		"/>
-        </g>
-        <g>
-          <line strokeWidth="1.5" fill="none" stroke="#AAAAAA" x1="17.4" y1="30.9" x2="25.5" y2="30.9"/>
-          <line strokeWidth="1.5" fill="none" stroke="#AAAAAA" x1="17.4" y1="86.4" x2="25.5" y2="86.4"/>
-          <line strokeWidth="1.5" fill="none" stroke="#AAAAAA" x1="9.3" y1="58.6" x2="25.5" y2="58.6"/>
-          <line strokeWidth="1.5" fill="none" stroke="#AAAAAA" x1="9.3" y1="114.2" x2="25.5" y2="114.2"/>
-          <line strokeWidth="1.5" fill="none" stroke="#AAAAAA" x1="17.4" y1="141.9" x2="25.5" y2="141.9"/>
-        </g>
-      </g>
-    </svg>
-  );
 };
 
 const formatMatrix = m => (
@@ -611,57 +306,60 @@ const serializeParams = (params) => ({
   x0: formatMatrix(JSON.stringify(params.x0))
 });
 
-const kalmanFilter = ({
-  inputs,
-  measurements,
-  A, B, C, D, P0, Q, R, x0
+//
+// Components
+//
+
+const Chart = ({
+  rows,
+  columnIndex,
+  outputs,
+  width
 }) => {
-  const Atrans = transpose(A);
-  const Ctrans = transpose(C);
-  const I = identity(x0.size());
+  const height = 200;
+  const rulerMarkSize = 40;
 
-  let y = measurements[0];
-  let x = x0;
-  let P = P0;
+  const samples = useMemo(
+    () => rows.map(r => Number(r[columnIndex])),
+  [rows, columnIndex]);
+  
+  const { min, max } = useMemo(() => samples.reduce(
+    ({ min, max }, sample) => ({
+      min: min === undefined ? sample : Math.min(min, sample),
+      max: max === undefined ? sample : Math.max(max, sample)
+    }), {}),
+  [samples]);
 
-  return measurements.map((z, index) => {
-    // Input
-    const u = inputs[index];
+  if (!width) return null;
 
-    // Update covariance
-    P = add(multiply(multiply(A, P), Atrans), Q);
-
-    // Optimize gain
-    const K = divide(
-      multiply(P, Ctrans),
-      add(
-        multiply(multiply(C, P), Ctrans),
-        R
-      ));
-
-    // Correct state with measurement
-    x = add(x, multiply(K, (z - y)));
-
-    // Correct covariance
-    P = add(
-      multiply(
-        multiply(subtract(I, multiply(K, C)), P),
-        transpose(subtract(I, multiply(K, C)))),
-      multiply(multiply(K, R), transpose(K))
-    );
-
-    // Predict
-    y = add(multiply(C, x), multiply(D, u))._data[0][0];
-
-    // Update state
-    x = add(multiply(A, x), multiply(B, u));
-
-    // Output
-    return y;
-  });
+  return (
+    <svg
+      width={`${width}px`}
+      height={`${height}px`}
+      viewBox={`0 0 ${width} ${height}`}
+    >
+      <ChartBackground />
+      <polyline
+        fill="none"
+        stroke="#FF008E"
+        strokeMiterlimit="10"
+        points={getPoints(
+          samples, min, max, width - rulerMarkSize, height - rulerMarkSize, rulerMarkSize, 0
+        )}
+      />
+      <polyline
+        fill="none"
+        stroke="#12C0E1"
+        strokeMiterlimit="10"
+        points={getPoints(
+          outputs, min, max, width - rulerMarkSize, height - rulerMarkSize, rulerMarkSize, 0
+        )}
+      />
+    </svg>
+  );
 };
 
-const InteractiveKalmanFilter = () => {
+const KalmanDemo = () => {
   const [ params, setParams ] = useState(serializeParams(defaultParams));
   const [ columnMap, setColumnMap ] = useState(defaultMapping);
   const [ columns, setColumns ] = useState(defaultInput[0]);
@@ -838,4 +536,4 @@ const InteractiveKalmanFilter = () => {
   );
 };
 
-export default InteractiveKalmanFilter;
+export default KalmanDemo;
