+++ src/articles/components/load.tsx
@@ -1,0 +1,42 @@
+import React from 'react';
+
+const LoadImage = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="65.4px"
+    height="81.7px"
+    viewBox="0 0 102.2 127.6"
+  >
+    <path fill="#CDCDCD" d="M0.7,39c0-6.6,0-13.1,0-19.7c0-3.2-0.2-6.4,0-9.5c0.1-2.7,1.5-5.2,3.7-6.7c2.7-1.7,5.4-2,8.5-1.9
+      c3.1,0.1,6.2,0.1,9.3,0.1c6.4-0.1,12.8,0,19.3-0.1c12.8-0.2,25.7,0,38.6,0c3.2,0,6.5,0,9.7,0c2.8,0,5.5,0,7.8,1.6
+      c2,1.4,3.5,3.4,3.9,5.9c0.2,1.5,0.1,3,0.1,4.5c0,1.7,0,3.3,0,5c0,13,0.3,26,0.3,39c0,3.1,0,6.1,0,9.2c0,0.7,0.4,3.6-1,3
+      c-1.4-0.6-2.7-1.6-4-2.4c-0.7-0.4-1.3-0.8-2-1.2c-0.5-0.3-1.4-0.8-1.8-1.3c-0.3-0.5-0.2-1.4-0.2-2c0-0.9-0.1-1.7-0.1-2.6
+      c-0.2-3.2-0.5-6.4-0.7-9.7c-0.1-1.5-0.1-3.1-0.4-4.6c-0.3-1.4-0.9-2.8-1.6-4c-0.7-1.2-1.5-2.2-2.4-3.1c-1-0.9-2-2-3.1-2.6
+      c-2.5-1.5-5.7-2-8.6-2c-3.2,0-6.5,0-9.7,0c-12.8,0-25.7-0.2-38.5,0.1c-3.1,0.1-5.9,0-8.8,1.3c-2.6,1.2-4.8,3.1-6.3,5.5
+      c-1.7,2.6-2.2,5.5-2.4,8.6c-0.2,3.2-0.3,6.4-0.6,9.7c-0.1,1.5,0,3.2-0.2,4.7c-0.2,1.2-1.6,1.7-2.5,2.2c-1.3,0.8-2.6,1.6-3.9,2.4
+      c-0.4,0.3-1.3,1-1.8,1c-0.7,0-0.6-0.8-0.6-1.3c0-6.5,0.1-13.1,0-19.6C0.7,45.2,0.7,42.1,0.7,39C0.7,39,0.7,39,0.7,39z"/>
+    <path fill="#CDCDCD" d="M96,107.5c-0.4,3.3,0,6.7-0.1,10c0,3-1.3,5.9-4,7.4c-2.4,1.4-5.7,1.7-8.2,0.4c-2.6-1.3-4.3-3.9-4.5-6.8
+      c-0.3-3.7,0-7.4-0.2-11.1c-11.5-0.4-23,0-34.5-0.1c-5.7-0.1-11.4-0.1-17.1,0c-0.6,0-1.3,0-1.9,0c-0.6,0-1.4-0.1-1.9,0.1
+      c-0.7,0.3-0.4,1.9-0.4,2.5c0,2.9,0.2,5.8,0,8.7c-0.3,4.6-4.9,7.9-9.3,7.4c-2.2-0.2-4.4-1.4-5.8-3.2c-1.7-2.2-1.6-4.8-1.6-7.4
+      c0-1.4,0-2.9,0-4.3c0-0.7,0-1.3,0-2c0-0.4,0.1-1,0-1.4c-0.2-0.7-1.4-0.4-1.9-0.4c-0.6,0-1.3,0-1.9,0c-0.4,0-1.1,0.1-1.5-0.1
+      c-0.7-0.4-0.4-2.6-0.4-3.2c0-1.4,0-2.8,0-4.1c0.1-2.8,0-5.6,0-8.4c0-2.9,0-5.8-0.1-8.7c0-1.5,0-2.9,0-4.4c0-1-0.4-2.6,0.3-3.5
+      c0.7-0.9,2.2-1.5,3.2-2.1c1.2-0.7,2.3-1.4,3.5-2.1c1.2-0.7,2.4-1.5,3.6-2.2c1-0.6,2.2-1.1,2.4-2.4c0.2-1.4,0.1-2.9,0.1-4.4
+      c0-1.4,0.1-2.8,0.2-4.2c0.2-2.9,0.4-5.7,0.6-8.6c0.2-2.7,0.9-5.1,2.8-7.1c1.8-1.9,4.1-3,6.7-3.3c1.4-0.2,2.9-0.1,4.3,0
+      c1.4,0,2.8,0,4.2,0c2.9,0,5.9,0,8.8,0c5.7,0,11.4,0,17.1,0c5.9,0,11.7-0.1,17.6,0c2.6,0,5.1,0.6,7.2,2.2c1.8,1.4,3.6,4,3.9,6.3
+      c0.6,5.7,0.9,11.5,1.3,17.2c0.1,1-0.1,2.5,0.7,3.3c0.5,0.4,1.1,0.7,1.7,1c0.6,0.4,1.2,0.7,1.8,1.1c2.3,1.4,4.7,2.8,7,4.2
+      c0.5,0.3,1.3,0.7,1.6,1.1c0.4,0.4,0.3,1.1,0.4,1.6c0.1,1.4,0,2.8,0,4.2c0,2.8,0.1,5.6,0.1,8.5c0,2.9-0.2,5.7-0.1,8.6
+      c0.1,2.9,0,5.7-0.1,8.6c0,1.3-0.9,1.2-2,1.2C98.5,107.4,97.3,107.4,96,107.5z M72.8,66.9c2.9,0,5.7,0,8.6,0.1c0.8,0,2.3,0.3,2.4-0.8
+      c0.1-1.3-0.1-2.8-0.2-4.1c-0.2-2.9-0.4-5.7-0.6-8.5c-0.1-1.4-0.3-2.9-0.3-4.3c0-1.2-0.2-2.3-0.9-3.3c-0.6-0.9-1.4-1.7-2.4-2.2
+      c-1.1-0.5-2.4-0.5-3.6-0.5c-2.9,0-5.8,0-8.7,0c-11.7-0.1-23.4,0-35.1,0c-1.4,0-2.9,0-4.3,0c-1.2,0-2.6-0.1-3.8,0.3
+      c-2.5,0.8-3.7,3.2-4,5.6c-0.2,2.8-0.3,5.6-0.5,8.5c-0.1,1.5-0.2,3-0.3,4.5c-0.1,1.3-0.4,2.8-0.2,4c0.2,1,1.3,0.8,2.1,0.8
+      c1.4-0.1,2.8,0,4.2,0C28.1,66.9,67.1,66.8,72.8,66.9z M35.8,75.8c-1.2,0-2.4,0-3.6,0c-1,0-1.9,0-2.8,0.6c-1.9,1.3-1.3,4-1.3,5.9
+      c0,2.3,0,4.5,0,6.8c0,1.1,0,2.3,0,3.4c0,1-0.2,2.3,0.1,3.2c0.3,0.8,1.1,0.7,1.8,0.7c1.1,0,2.1,0,3.2,0c2.3,0,4.5,0,6.8,0
+      c4.5,0,8.9,0,13.4,0.1c4.4,0,8.8-0.1,13.2,0c2.1,0,4.4,0.3,6.5,0c1.1-0.2,1.3-0.4,1.3-1.4c0-1,0-2,0-3c0-4.4,0.1-8.9-0.1-13.3
+      c-0.1-1.7-1.5-2.9-3.2-3.1C69,75.6,40.9,75.8,35.8,75.8z M88,78.9c-1.7,0.2-3,0.5-4.4,1.7c-1.3,1.1-1.9,2.3-2.2,4
+      c-0.4,2.7,0.7,5.2,3,6.7c2.2,1.5,5.1,1.4,7.3-0.1c2.4-1.7,3.4-4.7,2.5-7.6C93.4,80.9,90.9,79,88,78.9z M20.9,85.9
+      c0.1-1.7-0.4-3-1.5-4.3c-1-1.1-2.2-1.9-3.6-2.3c-2.5-0.8-5.3,0.4-6.8,2.5c-1.8,2.4-1.6,5.9,0.3,8.2c0.8,1,1.8,1.7,2.9,2.2
+      c1.5,0.6,2.8,0.6,4.2,0C19,91.2,21.2,88.8,20.9,85.9z"/>
+  </svg>
+);
+
+export default LoadImage;
